refactor(: abstract into components, organise blog code
