+++ src/pages/blog/index.astro
@@ -3,7 +3,7 @@ import AppLayout from "../../layouts/app/AppLayout.astro";
 import PageLayout from "../../layouts/PageLayout.astro";
 import { SITE_TITLE } from "../../consts";
 import { getCollection } from "astro:content";
-import BlogFilter from "../../components/BlogFilter.svelte";
+import BlogPosts from "../../components/BlogPosts.svelte";
 
 // Load blog posts
 const allPosts = await getCollection("blog");
@@ -14,10 +14,10 @@ const allPosts = await getCollection("blog");
         <h1 class="text-3xl font-bold mb-6 text-gray-900 dark:text-gray-100">
             {"Christian's"} Blog
         </h1>
-        <p class="text-lg text-gray-700 dark:text-gray-300 mb-8">
+        <p class="text-lg text-gray-700 dark:text-gray-300 mb-6">
             Welcome to my public blog! Here you'll find blog posts on various topics, including
             technology, personal experiences, and more.
         </p>
-        <BlogFilter client:load posts={allPosts} />
+        <BlogPosts client:load posts={allPosts} />
     </PageLayout>
 </AppLayout>
